refactor(fe): migrate content from AssignmentAccordionNew to AssignmentAccordion
